Migrated `distance_between()` to `leads.data`. (#462)
