+++ leads/data_persistence/analyzer/utils.py
@@ -1,8 +1,5 @@
 from typing import Any as _Any
 
-from leads.data import dlat2meters, dlon2meters
-from .._computational import sqrt as _sqrt
-
 
 def time_invalid(o: _Any) -> bool:
     return not isinstance(o, int)
@@ -30,15 +27,3 @@ def longitude_invalid(o: _Any) -> bool:
 
 def latency_invalid(o: _Any) -> bool:
     return not isinstance(o, int | float)
-
-
-def distance_between(lat_0: float, lon_0: float, lat: float, lon: float) -> float:
-    """
-    Calculate the distance between two locations on the Earth.
-    :param lat_0: the latitude of the first location
-    :param lon_0: the longitude of the first location
-    :param lat: the latitude of the second location
-    :param lon: the longitude of the second location
-    :return:
-    """
-    return _sqrt(dlon2meters(lon - lon_0, .5 * (lat_0 + lat)) ** 2 + dlat2meters(lat - lat_0) ** 2)
